requirements.txt
geographiclib==2.1
geopy==2.4.1
lxml==6.0.2
piexif==1.1.3
pillow==12.0.0
pillow_heif==1.1.1
python-docx==1.2.0
typing_extensions==4.15.0
